% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCRM.R
\name{updateModel}
\alias{updateModel}
\title{updateModel}
\usage{
updateModel(target, prior, data)
}
\arguments{
\item{target}{target toxicity level}

\item{prior}{prior guess for each dose}

\item{data}{the data to update the model with}
}
\description{
This updates the tite-crm
}

